+++ crates/services/txpool_v2/src/tests/mocks.rs
@@ -2,13 +2,19 @@ use crate::{
     error::Error,
     ports::{
         AtomicView,
+        BlockImporter as BlockImporterTrait,
+        ConsensusParametersProvider,
         GasPriceProvider,
+        NotifyP2P,
+        P2PRequests,
+        P2PSubscriptions,
         TxPoolPersistentStorage,
         WasmChecker,
         WasmValidityError,
     },
     GasPrice,
 };
+use fuel_core_services::stream::BoxStream;
 use fuel_core_storage::{
     Mappable,
     PredicateStorageRequirements,
@@ -18,18 +24,22 @@ use fuel_core_storage::{
     StorageSize,
 };
 use fuel_core_types::{
+    blockchain::{
+        header::ConsensusParametersVersion,
+        SealedBlock,
+    },
     entities::{
-        coins::coin::{
-            Coin,
-            CompressedCoin,
-        },
+        coins::coin::CompressedCoin,
         relayer::message::Message,
     },
     fuel_tx::{
         BlobId,
         Bytes32,
+        ConsensusParameters,
         Contract,
         ContractId,
+        Transaction,
+        TxId,
         UtxoId,
     },
     fuel_types::Nonce,
@@ -37,13 +47,22 @@ use fuel_core_types::{
         BlobBytes,
         BlobData,
     },
+    services::{
+        block_importer::{
+            ImportResult,
+            SharedImportResult,
+        },
+        p2p::{
+            GossipData,
+            GossipsubMessageAcceptance,
+            GossipsubMessageInfo,
+            PeerId,
+        },
+    },
 };
 use std::{
     borrow::Cow,
-    collections::{
-        HashMap,
-        HashSet,
-    },
+    collections::HashMap,
     sync::{
         Arc,
         Mutex,
@@ -56,7 +75,6 @@ pub struct Data {
     pub contracts: HashMap<ContractId, Contract>,
     pub blobs: HashMap<BlobId, BlobBytes>,
     pub messages: HashMap<Nonce, Message>,
-    pub spent_messages: HashSet<Nonce>,
 }
 
 #[derive(Clone, Default)]
@@ -65,14 +83,6 @@ pub struct MockDb {
 }
 
 impl MockDb {
-    pub fn insert_coin(&self, coin: Coin) {
-        self.data
-            .lock()
-            .unwrap()
-            .coins
-            .insert(coin.utxo_id, coin.compress());
-    }
-
     pub fn insert_dummy_blob(&self, blob_id: BlobId) {
         self.data
             .lock()
@@ -87,10 +97,6 @@ impl MockDb {
             .unwrap()
             .messages
             .insert(*message.id(), message);
-    }
-
-    pub fn spend_message(&self, id: Nonce) {
-        self.data.lock().unwrap().spent_messages.insert(id);
     }
 }
 
@@ -207,15 +213,10 @@ impl MockTxPoolGasPrice {
             gas_price: Some(gas_price),
         }
     }
-
-    pub fn new_none() -> Self {
-        Self { gas_price: None }
-    }
 }
 
-#[async_trait::async_trait]
 impl GasPriceProvider for MockTxPoolGasPrice {
-    async fn next_gas_price(&self) -> Result<GasPrice, Error> {
+    fn next_gas_price(&self) -> Result<GasPrice, Error> {
         self.gas_price
             .ok_or(Error::GasPriceNotFound("Gas price not found".to_string()))
     }
@@ -237,5 +238,99 @@ impl WasmChecker for MockWasmChecker {
         _wasm_root: &Bytes32,
     ) -> Result<(), WasmValidityError> {
         self.result
+    }
+}
+
+mockall::mock! {
+    pub ConsensusParametersProvider {}
+
+    impl ConsensusParametersProvider for ConsensusParametersProvider {
+        fn latest_consensus_parameters(&self) -> (ConsensusParametersVersion, Arc<ConsensusParameters>);
+    }
+}
+
+type GossipedTransaction = GossipData<Transaction>;
+
+mockall::mock! {
+    pub P2P {}
+
+    impl P2PSubscriptions for P2P {
+        type GossipedTransaction = GossipedTransaction;
+
+        fn gossiped_transaction_events(&self) -> BoxStream<GossipedTransaction>;
+
+        fn subscribe_new_peers(&self) -> BoxStream<PeerId>;
+    }
+
+    impl NotifyP2P for P2P {
+        fn notify_gossip_transaction_validity(
+            &self,
+            message_info: GossipsubMessageInfo,
+            validity: GossipsubMessageAcceptance,
+        ) -> anyhow::Result<()>;
+
+        fn broadcast_transaction(&self, transaction: Arc<Transaction>) -> anyhow::Result<()>;
+    }
+
+    #[async_trait::async_trait]
+    impl P2PRequests for P2P {
+        async fn request_tx_ids(&self, peer_id: PeerId) -> anyhow::Result<Vec<TxId>>;
+
+        async fn request_txs(
+            &self,
+            peer_id: PeerId,
+            tx_ids: Vec<TxId>,
+        ) -> anyhow::Result<Vec<Option<Transaction>>>;
+    }
+}
+
+impl MockP2P {
+    pub fn new_with_txs(txs: Vec<Transaction>) -> Self {
+        let mut p2p = MockP2P::default();
+        p2p.expect_gossiped_transaction_events().returning(move || {
+            let txs_clone = txs.clone();
+            let stream = fuel_core_services::stream::unfold(txs_clone, |mut txs| async {
+                let tx = txs.pop();
+                if let Some(tx) = tx {
+                    Some((GossipData::new(tx, vec![], vec![]), txs))
+                } else {
+                    core::future::pending().await
+                }
+            });
+            Box::pin(stream)
+        });
+
+        p2p
+    }
+}
+
+mockall::mock! {
+    pub Importer {}
+
+    impl BlockImporterTrait for Importer {
+        fn block_events(&self) -> BoxStream<SharedImportResult>;
+    }
+}
+
+impl MockImporter {
+    pub fn with_blocks(blocks: Vec<SealedBlock>) -> Self {
+        let mut importer = MockImporter::default();
+        importer.expect_block_events().returning(move || {
+            let blocks = blocks.clone();
+            let stream = fuel_core_services::stream::unfold(blocks, |mut blocks| async {
+                let block = blocks.pop();
+                if let Some(sealed_block) = block {
+                    let result: SharedImportResult = Arc::new(
+                        ImportResult::new_from_local(sealed_block, vec![], vec![]),
+                    );
+
+                    Some((result, blocks))
+                } else {
+                    core::future::pending().await
+                }
+            });
+            Box::pin(stream)
+        });
+        importer
     }
 }